+++ src/Lib.hs
@@ -14,8 +14,8 @@ module Lib
     ,getValOnByDate,sumValTs,subTsBetweenDates,splitTsByDate
     ,paySeqLiabilitiesAmt,getIntervalDays,getIntervalFactors
     ,zipWith8,zipWith9,zipWith10,zipWith11,zipWith12
-    ,weightedBy, mkTs, DealStatus(..)
-    ,mkRateTs,Pre(..)
+    ,weightedBy, mkTs
+    ,mkRateTs
     ) where
 
 import qualified Data.Time as T
@@ -28,14 +28,20 @@ import Data.Aeson.TH
 import Data.Aeson.Types
 import Data.Aeson hiding (json)
 import Text.Regex.TDFA
-import Data.Fixed
+import Data.Fixed (Fixed(..), HasResolution,Centi, resolution)
+import Data.Ratio
 import Types
 import Control.Lens
 import Data.List.Lens
 import Control.Lens.TH
+-- import Deal.DealType
+
 
 import Debug.Trace
 debug = flip trace
+
+
+
 
 annualRateToPeriodRate :: Period -> Float -> Float
 annualRateToPeriodRate p annualRate =